llm-doc-processor/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
numpy
scikit-learn
python-docx
PyPDF2
tqdm
sentence-transformers
faiss-cpu
requests
pandas
pytest
httpx
jinja2
loguru
spacy
google-generativeai
PyMuPDF
